third_party/blink/web_tests/platform/generic/webexposed/event-target-in-prototype-expected.txt

Check that interfaces inherit from EventTarget

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS EventSource.prototype.__proto__ is EventTarget.prototype
PASS FileReader.prototype.__proto__ is EventTarget.prototype
PASS MediaSource.prototype.__proto__ is EventTarget.prototype
PASS MessagePort.prototype.__proto__ is EventTarget.prototype
PASS Node.prototype.__proto__ is EventTarget.prototype
PASS Performance.prototype.__proto__ is EventTarget.prototype
PASS SharedWorker.prototype.__proto__ is EventTarget.prototype
PASS WebSocket.prototype.__proto__ is EventTarget.prototype
PASS Worker.prototype.__proto__ is EventTarget.prototype
PASS Window.prototype.__proto__.__proto__ is EventTarget.prototype
PASS successfullyParsed is true

TEST COMPLETE
